requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
cohere>=5.5.6
langchain-cohere>=0.3.4
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-magic==0.4.27